requirements.txt
kubernetes==9.0.0
PyYAML
